+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Issue introduced in v1.1.0 that caused a partial BAM file to be output.
 - Corrected example command in README.
 - Incorrect reporting of unique gene and transcripts in report table.
+- Processed expression matrix entries incorrectly filtered.
 ### Changed
 - Read chunking done in library code.
 - `--process_chunk_size` parameter changed to `--fastq_chunk`